authn-service/requirements.txt
asn1crypto==0.24.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
pycrypto==2.6.1
PyGObject==3.30.4
PyJWT==1.7.1
python-dotenv==0.15.0
pyxdg==0.25
requests==2.24.0
SecretStorage==2.3.1
six==1.12.0
urllib3==1.25.11
Werkzeug==1.0.1
